pyproject.toml
[project]
name = "zfield"
version = "0.2.1"
description = "Serial terminal and device manager for Zephyr RTOS development"
requires-python = ">=3.10"
dependencies = [
    "pywebview>=4.0",  # Compatible with Python 3.10+
    "fastapi>=0.100.0",  # Compatible with Python 3.10+
    "uvicorn[standard]>=0.23.0",  # Compatible with Python 3.10+
    "pyserial>=3.5",
    "pydantic>=2.0.0",  # Pydantic v2 supports Python 3.10+
    "pydantic-settings>=2.0.0",  # Compatible with Python 3.10+
    "pyinstaller>=5.0",  # Compatible with Python 3.10+
    "psutil>=5.9.0",
]

[project.optional-dependencies]
windows = [
    "pywinstyles",  # For native Windows Mica/Acrylic glass effect
    "pywin32>=305",  # For Windows device notifications
    "WMI>=1.5.1",  # Alternative Windows device monitoring
]
linux = [
    "pywebview[qt]>=4.0",
    "PyQt6>=6.4.0",  # Compatible with Python 3.10+
    "PyQt6-WebEngine>=6.4.0",  # Compatible with Python 3.10+
    "pyudev>=0.24.0",  # For Linux udev device monitoring
    "inotify>=0.2.10",  # Fallback Linux file system monitoring
]
dev = [
    "pytest",
    "black",
    "ruff",
]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["backend.app"]
